Move some tests to integration tests
